Adding Files
This is all the code I did myself. If you add these to your own STM32IDE it will not work since I used STM32CUBEMX to configure the UART and ADC. All the .c files go into the Src folder. The .h files go into the Inc folder. To enable USART3, click into Testing.ioc. There should be a link on the left side of the STM32CubeIDE program. Click on the Connectivity tab on the left in STM32CUBEMX. Change the mode to Asynchronous. To enable the ADC go to the Analog tab on the left. Change IN1 and IN3 to Single-ended.