Adding SCD41 Code
This code gets values send on the data line. The values are not correct and nothing is being printed to the terminal, however I can see data being sent on the line. Also, the clock isn't running which is weird but I want to save the code in case I mess anything up.  

Configuration: PC8 as SCL, PB10 as SDA. Both are pulled up. I2C1 enabled for sensor. UART2 enabled for Putty.